derappp: Data for Environmental Risk Assessment of Plant Protection Products

Makes some public data for environmental risk assessment of plant protection products accessible in R. Includes definitions of substances in terms of well-defined chemical entities, if applicable. Also includes source references including page numbers, most often referring to secondary sources, as the original reports are usually unpublished. Uses standardised scientific names for biological species and includes physicochemical as well as ecotoxicological data. Data sources are mainly conclusions from the European pesticide risk assessment peer review process published by the European Food Safety Authority (EFSA) and Renewal Assessment Reports (RARs), also published by EFSA. Note that the use of the data included in this package for registration purposes may be restricted by regulatory data protection, also known as data exclusivity rules, as detailed for example in article 59 of Regulation (EC) No 1107/2009.

See also

Author

Maintainer: Johannes Ranke johannes.ranke@agroscope.admin.ch (ORCID) (acronym: rajo)

Authors:

Johannes Ranke johannes.ranke@agroscope.admin.ch (ORCID) (acronym: rajo)

Elisabeth Lutz elisabeth.lutz@agroscope.admin.ch (acronym: luel, data generation scripts, aquatic toxicity data)

Marcel Mathis marcel.mathis@agroscope.admin.ch (acronym: mamr, data generation scrips, aquatic toxicity data)

Other contributors:

Laurent Boualit (acronym: bolr, aquatic toxicity data) [contributor]

Jhoniel Perdigon Ferreira (acronym: pejh, data generation scripts, soil toxicity data, review of aquatic toxicity data) [contributor]

Kasteel Romualdus romualdus.kasteel@agroscope.admin.ch (Reviewed the initial version before publication) [reviewer]

Daniel Baumgartner (SMILES codes for pyrethroid stereoisomers) [contributor]

Fabian Balk (ORCID) (acronym: bafa, review of soil toxicity data) [reviewer]

European Food Safety Authority (EFSA conclusions, bibliographic information in OpenFoodTox, Renewal Assessment Reports) [data contributor]

British Crop Protection Council (Chemical identity information from the Compendium of Pesticide Common Names) [data contributor]

National Center for Biotechnology Information (Chemical identity information retrieved from the PubChem website) [data contributor]

CAS division of the American Chemical Society (Chemical identity information retreived from the CAS Common Chemistry website) [data contributor]

enviPath UG & Co KG (Chemical identity information retreived from the envipath website) [data contributor]

Agroscope [copyright holder]